+++ PSTibber.psm1
@@ -1,7 +1,17 @@
 ﻿$functionsRoot = Join-Path $PSScriptRoot functions
-foreach ($file in Get-ChildItem -Path "$functionsRoot" -Filter *-*.ps1 -Recurse) {
-    . $file.FullName
+$publicFunctions = @( Get-ChildItem -Path "$functionsRoot\*.ps1" -Recurse -ErrorAction SilentlyContinue | `
+        Where-Object { $_.FullName -notlike "$functionsRoot\internal\*" } )
+$privateFunctions = @( Get-ChildItem -Path "$functionsRoot\internal\*.ps1" -Recurse -ErrorAction SilentlyContinue )
+
+$allFunctions = $publicFunctions + $privateFunctions
+foreach ($function in $AllFunctions) {
+    try {
+        . $function.Fullname
+    }
+    catch {
+        Write-Error -Message "Failed to import function $($function.Fullname): $_"
+    }
 }
 
-# Export all aliases and verb-noun functions
-Export-ModuleMember -Alias * -Function *-*
+# Export the public functions
+Export-ModuleMember -Function $publicFunctions.BaseName