+++ functions/Register-TibberLiveConsumptionSubscription.ps1
@@ -53,7 +53,13 @@
             'lastMeterProduction'
             'powerFactor'
             'signalStrength'
-        )
+        ),
+
+        # Specifies the time to wait for WebSocket operations, or -1 to wait indefinitely.
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [ValidateRange(-1, [int]::MaxValue)]
+        [Alias('Timeout')]
+        [int] $TimeoutInSeconds = 10
     )
 
     begin {
@@ -73,22 +79,8 @@
                 query = "subscription{ liveMeasurement(homeId:`"$homeId`"){ $($Fields -join ','),__typename }}"
             }
         } | ConvertTo-Json -Depth 10
-        Write-Debug -Message ("Invoking GraphQL over WebSocket command: " + $uri)
-        Write-Debug -Message "GraphQL over WebSocket command:"
-        Write-Debug -Message $command
-        $sendCommand = New-Object ArraySegment[byte] -ArgumentList @(, $([Text.Encoding]::ASCII.GetBytes($command)))
-        $result = $webSocket.SendAsync($sendCommand, [Net.WebSockets.WebSocketMessageType]::Text, $true, $cancellationToken)
-        while (-Not $result.IsCompleted) {
-            Start-Sleep -Milliseconds 10
-        }
-        Write-Debug -Message "WebSocket status:"
-        Write-Debug -Message ($webSocket | Select-Object * | Out-String)
-        Write-Debug -Message "WebSocket operation result:"
-        Write-Debug -Message ($result | Select-Object * | Out-String)
-        if ($result.Result.CloseStatus) {
-            throw "Send failed: $($result.Result.CloseStatusDescription) [$($result.Result.CloseStatus)]"
-        }
-        Write-Verbose "Subscribe request sent"
+        Write-WebSocket -Data $command -WebSocket $webSocket -CancellationToken $cancellationToken -TimeoutInSeconds $TimeoutInSeconds
+        Write-Verbose "Subscribe request sent to: $uri [subscribe]"
 
         # Output subscription object
         [PSCustomObject]@{