+++ functions/Unregister-TibberLiveConsumptionSubscription.ps1
@@ -20,7 +20,13 @@
 
         # Specifies the subscripton to manage.
         [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName)]
-        [Object] $Subscription
+        [Object] $Subscription,
+
+        # Specifies the time to wait for WebSocket operations, or -1 to wait indefinitely.
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [ValidateRange(-1, [int]::MaxValue)]
+        [Alias('Timeout')]
+        [int] $TimeoutInSeconds = 10
     )
 
     begin {
@@ -37,21 +43,7 @@
             id   = $subscriptionId
             type = 'stop'
         } | ConvertTo-Json -Depth 10
-        Write-Debug -Message ("Invoking GraphQL over WebSocket command: " + $uri)
-        Write-Debug -Message "GraphQL over WebSocket command:"
-        Write-Debug -Message $command
-        $sendCommand = New-Object ArraySegment[byte] -ArgumentList @(, $([Text.Encoding]::ASCII.GetBytes($command)))
-        $result = $webSocket.SendAsync($sendCommand, [Net.WebSockets.WebSocketMessageType]::Text, $true, $cancellationToken)
-        while (-Not $result.IsCompleted) {
-            Start-Sleep -Milliseconds 10
-        }
-        Write-Debug -Message "WebSocket status:"
-        Write-Debug -Message ($webSocket | Select-Object * | Out-String)
-        Write-Debug -Message "WebSocket operation result:"
-        Write-Debug -Message ($result | Select-Object * | Out-String)
-        if ($result.Result.CloseStatus) {
-            throw "Send failed: $($result.Result.CloseStatusDescription) [$($result.Result.CloseStatus)]"
-        }
-        Write-Verbose "Unsubscribe (stop) request sent"
+        Write-WebSocket -Data $command -WebSocket $webSocket -CancellationToken $cancellationToken -TimeoutInSeconds $TimeoutInSeconds
+        Write-Verbose "Unsubscribe request sent to: $uri [stop]"
     }
 }